Remove duplication from volunteering and work experience forms
